ui: port Rad Racer 8-bit theme to the comma four (MICI)
Brings the BPRadRacerTheme onroad scene (black sky, star field, parallax
skyline, roadside signs, green game road with red/white curbs and
scrolling dashes, ego + lead car sprites) to the 536x240 MICI display,
with an on/off toggle ("8-Bit Racer Theme") in the MICI visuals
settings. Param default is off; with the toggle off every renderer
follows its existing code path unchanged.

Structure:
- The road-drawing methods and constants move verbatim from the TICI
  ModelRendererBP into a shared RadRacerRoadMixin (rad_racer_road.py);
  the TICI class now inherits the mixin. No TICI value or behavior
  changes.
- RadRacerTheme's screen-scale knobs (skyline scale, star count, sprite
  scales, gaps) become class attributes -- TICI values unchanged --
  so RadRacerThemeMici can subclass with ~0.25x MICI scales.
- No gauge cluster on MICI: at 240 px tall it would cover the whole
  screen, and the MICI HUD/complication already shows that data. The
  ego sprite anchors to the bottom edge instead.
- MICI wiring mirrors TICI: camera blacks out via the shared
  _should_hide_camera_view gate, background renders after the camera
  and before the model road, sprites render after; params cached and
  refreshed on the existing ~1 s cadences.

Known MICI-specific behavior: the base MICI renderer skips all road
drawing while DISENGAGED, so the theme shows sky + ego only until
engagement (TICI shows a dim gray road); consistent with MICI's
existing minimal-view behavior.
